refactor: re-organize flags packages so graph is cleaner (#781)
Pre-factor to finishing work for the CLI MVP

GitOrigin-RevId: 4068f4fd34cfca54853b907d2e962eee807655bd
