+++ cmd/docgen/BUILD.bazel
@@ -7,7 +7,6 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/aspect/root",
-        "//pkg/aspect/root/flags",
         "//pkg/bazel",
         "//pkg/ioutils",
         "//pkg/plugin/system",
