+++ cmd/docgen/main.go
@@ -26,7 +26,6 @@ import (
 	"github.com/spf13/cobra/doc"
 
 	"aspect.build/cli/cmd/aspect/root"
-	"aspect.build/cli/pkg/aspect/root/flags"
 	"aspect.build/cli/pkg/bazel"
 	"aspect.build/cli/pkg/ioutils"
 	"aspect.build/cli/pkg/plugin/system"
@@ -55,7 +54,7 @@ func main() {
 	aspectRootCmd := root.NewDefaultRootCmd(pluginSystem)
 
 	// Run this command after all bazel verbs have been added to "cmd".
-	if err := flags.AddBazelFlags(cmd); err != nil {
+	if err := bazel.AddBazelFlags(cmd); err != nil {
 		log.Fatal(err)
 	}
